New README and removing automatic stopping
The README was outdated, so it has been updated. The project structure has also been moved so the configuration is at the top, allowing for easier access without having to scroll through many lines of code first.

Automatic stopping  didn't work very well from my experience, so it has been removed. The regions for screenshots will likely be updated later so it is not captured.